+++ src/Desktop/Portal/Request.hs
@@ -1,3 +1,0 @@
-module Desktop.Portal.Request (module Desktop.Portal.Request.Internal) where
-
-import Desktop.Portal.Request.Internal (Client, Request, await, cancel, connect, disconnect)