+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ## 0.1.0.0
 ### Added
 - Make client connections explicit so they can be shared across requests and signal handlers.
+- Add Notification portal support.
 
 ## 0.0.1.0